Implement asynchronous streaming response for chat interactions and update frontend title during streaming
