+++ backend/services/__init__.py
@@ -1,6 +1,5 @@
 from services.llm_agent import (
     create_agent,
-    stream_agent_response,
 )
 from services.tools_initializer import (
     create_query_engines_from_filters,
